Fix: add CollectorConfig, global config for collector classes
