+++ dspot/src/main/java/eu/stamp_project/Main.java
@@ -1,5 +1,6 @@
 package eu.stamp_project;
 
+import eu.stamp_project.utils.collector.CollectorConfig;
 import eu.stamp_project.dspot.DSpot;
 import eu.stamp_project.utils.options.JSAPOptions;
 import eu.stamp_project.utils.program.InputConfiguration;
@@ -60,7 +61,7 @@ public class Main {
 		Main.GLOBAL_REPORT.output();
 		Main.GLOBAL_REPORT.reset();
 		// Send info collected.
-		InputConfiguration.getInformationCollector().sendInfo();
+		CollectorConfig.getInstance().getInformationCollector().sendInfo();
 	}
 
 	public static void createOutputDirectories() {
