+++ dspot/src/main/java/eu/stamp_project/dspot/selector/JacocoCoverageSelector.java
@@ -8,12 +8,14 @@ import eu.stamp_project.utils.report.output.selector.coverage.json.TestCaseJSON;
 import eu.stamp_project.utils.report.output.selector.coverage.json.TestClassJSON;
 import eu.stamp_project.testrunner.listener.Coverage;
 import eu.stamp_project.testrunner.listener.CoveragePerTestMethod;
+import eu.stamp_project.utils.collector.CollectorConfig;
 import eu.stamp_project.utils.program.InputConfiguration;
 import eu.stamp_project.testrunner.EntryPoint;
 import eu.stamp_project.utils.AmplificationHelper;
 import eu.stamp_project.utils.Counter;
 import eu.stamp_project.utils.DSpotUtils;
 import eu.stamp_project.utils.compilation.DSpotCompiler;
+
 import org.apache.commons.io.FileUtils;
 import spoon.reflect.declaration.CtMethod;
 import spoon.reflect.declaration.CtNamedElement;
@@ -259,7 +261,7 @@ public class JacocoCoverageSelector extends TakeAllSelector {
                 )
         );
 
-        InputConfiguration.getInformationCollector().reportSelectorInformation(testClassJSON.toString());
+        CollectorConfig.getInstance().getInformationCollector().reportSelectorInformation(testClassJSON.toString());
         return testClassJSON;
     }
 }
