+++ dspot/src/main/java/eu/stamp_project/dspot/selector/PitMutantScoreSelector.java
@@ -18,6 +18,7 @@ import eu.stamp_project.utils.report.output.selector.mutant.json.TestCaseJSON;
 import eu.stamp_project.utils.report.output.selector.mutant.json.TestClassJSON;
 import eu.stamp_project.utils.Counter;
 import eu.stamp_project.utils.DSpotUtils;
+import eu.stamp_project.utils.collector.CollectorConfig;
 import eu.stamp_project.utils.program.InputConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -311,7 +312,7 @@ public class PitMutantScoreSelector extends TakeAllSelector {
                     }
                 }
         );
-        InputConfiguration.getInformationCollector().reportSelectorInformation(testClassJSON.toString());
+        CollectorConfig.getInstance().getInformationCollector().reportSelectorInformation(testClassJSON.toString());
         return testClassJSON;
     }
